These are all the files in my Visual Studio Solution/project
This project is a Specflow (cucumber for .NET) project, based on BDD or
gherkin syntax. Essentially, the selenium script navigates to the Moby
Max registration page, fills out the appropriate fields and controls
with data, makes the form submission, and checks that the resultant
welcome page is displayed. More negative and field validation tests may
be forthcoming.